simplify orchestration and entity context interfaces and implementation
